requirements.txt
python-telegram-bot==11.1.0
requests==2.25.1
python-dotenv==0.15.0
google-cloud-dialogflow==2.1.2
google-cloud-storage==1.37.1
vk_api==11.9.1
